+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Backwards incompatible changes
     - Minimum required elixir version is now `~> 1.7`
 
+- Bug fixes
+    - All cache adapter callbacks now receive the database configuration with the `:id` properly set
+
 ## v0.1.0 (2019-10-25)
 
 - Initial Release